+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Raise an error when an ignore file does not parse to a hash. ([@bobbymcwho][])
 - Log all filtered backtrace lines to the logger ([@bobbymcwho][])
 - Add support for removing dynamic adapters. ([@Mange][])
+- Allow aliases in .isolator_todo.yml and .isolator_ignore.yml ([@tomgi][])
 
 ## 0.8.0 (2021-12-29)